requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
librosa
pydub
scikit-learn
numpy
joblibpython-dotenv==1.1.1
